+++ frontend/tsconfig.paths.json
@@ -6,7 +6,6 @@
 			"@assets/*": ["assets/*"],
       "@components/*": ["components/*"],
 			"@constants/*": ["constants/*"],
-			"@functions/*": ["functions/*"],
 			"@hooks/*": ["hooks/*"],
 			"@libs/*": ["libs/*"],
 			"@mocks/*": ["mocks/*"],
